third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_lostpointercapture_for_disconnected_node-expected.txt

This is a testharness.js-based test.
[FAIL] lostpointercapture must be received before the next pointerevent after the node is disconnected
  assert_equals: No pointerevents should be received before lost capture is resolved expected "" but got "pointerleave@BODY, pointermove@target0"
Harness: the test ran to completion.
